+++ lib/screens/document_type_screen.dart
@@ -1,4 +1,5 @@
 import 'package:adaptive_dialog/adaptive_dialog.dart';
+import 'package:connectivity/connectivity.dart';
 import 'package:flutter/material.dart';
 
 import 'package:vehicles_app/components/loader_component.dart';
@@ -149,6 +150,24 @@ class _DocumentTypeScreenState extends State<DocumentTypeScreen> {
       _showLoader = true;
     });
 
+    var connectivityResult = await Connectivity().checkConnectivity();
+
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+
+      await showAlertDialog(
+        context: context,
+        title: 'Error',
+        message: 'Verifica que estes conectado a internet',
+        actions: <AlertDialogAction>[
+          AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );
+      return;
+    }
+
     Map<String, dynamic> request = {
       'description': _description,
     };
@@ -156,7 +175,7 @@ class _DocumentTypeScreenState extends State<DocumentTypeScreen> {
     Response response = await ApiHelper.post(
       '/api/DocumentTypes/', 
       request, 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
@@ -183,6 +202,24 @@ class _DocumentTypeScreenState extends State<DocumentTypeScreen> {
       _showLoader = true;
     });
 
+    var connectivityResult = await Connectivity().checkConnectivity();
+
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+
+      await showAlertDialog(
+        context: context,
+        title: 'Error',
+        message: 'Verifica que estes conectado a internet',
+        actions: <AlertDialogAction>[
+          AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );
+      return;
+    }
+
     Map<String, dynamic> request = {
       'id': widget.documentType.id,
       'description': _description,
@@ -192,7 +229,7 @@ class _DocumentTypeScreenState extends State<DocumentTypeScreen> {
       '/api/DocumentTypes/', 
       widget.documentType.id.toString(), 
       request, 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
@@ -235,10 +272,28 @@ class _DocumentTypeScreenState extends State<DocumentTypeScreen> {
       _showLoader = true;
     });
 
+    var connectivityResult = await Connectivity().checkConnectivity();
+
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+
+      await showAlertDialog(
+        context: context,
+        title: 'Error',
+        message: 'Verifica que estes conectado a internet',
+        actions: <AlertDialogAction>[
+          AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );
+      return;
+    }
+
     Response response = await ApiHelper.delete(
       '/api/DocumentTypes/', 
       widget.documentType.id.toString(), 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
