+++ lib/screens/procedure_screen.dart
@@ -1,4 +1,5 @@
 import 'package:adaptive_dialog/adaptive_dialog.dart';
+import 'package:connectivity/connectivity.dart';
 import 'package:flutter/material.dart';
 
 import 'package:vehicles_app/components/loader_component.dart';
@@ -200,6 +201,24 @@ class _ProcedureScreenState extends State<ProcedureScreen> {
       _showLoader = true;
     });
 
+    var connectivityResult = await Connectivity().checkConnectivity();
+
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+
+      await showAlertDialog(
+        context: context,
+        title: 'Error',
+        message: 'Verifica que estes conectado a internet',
+        actions: <AlertDialogAction>[
+          AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );
+      return;
+    }
+
     Map<String, dynamic> request = {
       'description': _description,
       'price': double.parse(_price),
@@ -208,7 +227,7 @@ class _ProcedureScreenState extends State<ProcedureScreen> {
     Response response = await ApiHelper.post(
       '/api/Procedures/', 
       request, 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
@@ -235,6 +254,24 @@ class _ProcedureScreenState extends State<ProcedureScreen> {
       _showLoader = true;
     });
 
+    var connectivityResult = await Connectivity().checkConnectivity();
+
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+
+      await showAlertDialog(
+        context: context,
+        title: 'Error',
+        message: 'Verifica que estes conectado a internet',
+        actions: <AlertDialogAction>[
+          AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );
+      return;
+    }
+
     Map<String, dynamic> request = {
       'id': widget.procedure.id,
       'description': _description,
@@ -245,7 +282,7 @@ class _ProcedureScreenState extends State<ProcedureScreen> {
       '/api/Procedures/', 
       widget.procedure.id.toString(), 
       request, 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
@@ -288,10 +325,28 @@ class _ProcedureScreenState extends State<ProcedureScreen> {
       _showLoader = true;
     });
 
+    var connectivityResult = await Connectivity().checkConnectivity();
+
+    if (connectivityResult == ConnectivityResult.none) {
+      setState(() {
+        _showLoader = false;
+      });
+
+      await showAlertDialog(
+        context: context,
+        title: 'Error',
+        message: 'Verifica que estes conectado a internet',
+        actions: <AlertDialogAction>[
+          AlertDialogAction(key: null, label: 'Aceptar'),
+        ]
+      );
+      return;
+    }
+
     Response response = await ApiHelper.delete(
       '/api/Procedures/', 
       widget.procedure.id.toString(), 
-      widget.token.token
+      widget.token
     );
 
     setState(() {
